back_fastapi/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
psycopg2-binary
python-dotenv
requests